Mock_Trial_Demo/requirements.txt
flask
flask_jwt_extended
Flask-SQLAlchemy
flask-login
openai==0.27.6
scikit-learn
sentence-transformers
keyboard
PyPDF2
pdfminer.six
pdfplumber
PyMuPDF
textract
gspread==5.9.0
oauth2client==4.1.3
huggingface_hub==0.16.4
wrapt==1.15.0
google-cloud-texttospeech
Werkzeug==2.3.7
git+https://github.com/huggingface/transformers.git
accelerate
soundfile==0.12.1 
speechbrain==0.5.15
sentencepiece==0.1.99
spacy== 3.7.2
numpy==1.22.4
regex==2023.10.3




